verdant_query_preprocessor

Function process_locals

pub fn process_locals(strip_comment: &str, nvim_like: bool, src: &str) -> String

Pre-process locals queries.

See the crate docs for information on the parameters.

Panics

The function panics if the query string cannot be parsed by rsexpr.